requirements.txt
# Core UI & App
streamlit
streamlit-autorefresh
python-dotenv
requests
pandas
numpy
matplotlib
plotly
Pillow
ta

# PDF & Report Generation
fpdf

# Market Data
yfinance

# Bybit SDK
pybit

# Reddit API
praw

# Database
sqlalchemy
psycopg2-binary

# Machine Learning
scikit-learn
joblib
xgboost